Extract queue drag-to-reorder into its own composable
Move the pointer-driven reorder logic out of useFullscreenQueue (which was growing large again) into a focused, independently testable useQueueDragReorder composable. No behaviour change.
